built_graph update
